CW_7 - created Header for User Element, my profile page created
