+++ overAll/src/main/java/org/pages/HomePage.java
@@ -1,40 +1,24 @@
 package org.pages;
 
-import org.apache.log4j.Logger;
-
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
+import org.pages.elements.HeaderForUserElement;
 
 public class HomePage extends ParentPage{
 //    private Logger logger = Logger.getLogger(getClass());
 
-    @FindBy(xpath = "//button[text()='Sign Out']")
-    private WebElement buttonSignOut;
-
-    @FindBy(xpath = "//a[@class='btn btn-sm btn-success mr-2']")
-    private WebElement buttonCreatePost;
 
     public HomePage(WebDriver webDriver) {
         super(webDriver);
     }
 
-
-    public void checkIsButtonSignOutVisible() {
-//        Assert.assertTrue("Button Sign Out is not visible" , isButtonSighOutVisible());
-        checkIsElementVisible(buttonSignOut);
+    public HeaderForUserElement getHeaderElement() {
+        return new HeaderForUserElement(webDriver);
     }
 
     public HomePage checkIsRedirectToHomePage() {
-        checkIsButtonSignOutVisible();
+        getHeaderElement().checkIsButtonSignOutVisible();
         //TODO check current URL
         return this;
     }
-
-    public CreateNewPostPage clickOnButtonCreatePost() {
-        clickOnElement(buttonCreatePost);
-        return new CreateNewPostPage(webDriver);
-    }
-
 
 }
