+++ overAll/src/main/java/org/pages/PageProvider.java
@@ -17,4 +17,6 @@ public class PageProvider {
         return new HomePage(webDriver);
     }
 
+    public PostPage getPostPage() {return new PostPage(webDriver);    }
+
 }
