+++ overAll/src/main/java/org/pages/PostPage.java
@@ -3,6 +3,7 @@ package org.pages;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+import org.pages.elements.HeaderForUserElement;
 
 public class PostPage extends ParentPage {
     // "//*[contains(@class,'alert-success')]"
@@ -11,6 +12,10 @@ public class PostPage extends ParentPage {
 
     public PostPage(WebDriver webDriver) {
         super(webDriver);
+    }
+
+    public HeaderForUserElement getHeaderForUserElement() {
+        return new HeaderForUserElement(webDriver);
     }
 
     public PostPage checkIsRedirectToPostPage() {
