+++ overAll/src/main/java/org/pages/elements/HeaderForUserElement.java
@@ -1,0 +1,37 @@
+package org.pages.elements;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.pages.CommonActionsWithElements;
+import org.pages.CreateNewPostPage;
+import org.pages.MyProfilePage;
+
+public class HeaderForUserElement extends CommonActionsWithElements {
+    @FindBy(xpath = "//img[@alt='My profile']")
+    private WebElement buttonMyProfile;
+
+    @FindBy(xpath = "//button[text()='Sign Out']")
+    private WebElement buttonSignOut;
+
+    @FindBy(xpath = "//a[@class='btn btn-sm btn-success mr-2']")
+    private WebElement buttonCreatePost;
+
+    public HeaderForUserElement(WebDriver webDriver) {
+        super(webDriver);
+    }
+    public MyProfilePage clickOnButtonMyProfile() {
+        clickOnElement(buttonMyProfile);
+        return new MyProfilePage(webDriver);
+    }
+    public void checkIsButtonSignOutVisible() {
+//        Assert.assertTrue("Button Sign Out is not visible" , isButtonSighOutVisible());
+        checkIsElementVisible(buttonSignOut);
+    }
+
+    public CreateNewPostPage clickOnButtonCreatePost() {
+        clickOnElement(buttonCreatePost);
+        return new CreateNewPostPage(webDriver);
+    }
+
+}
