+++ overAll/src/test/java/org/loginTest/LoginTestWithPageObject.java
@@ -15,6 +15,6 @@ public class LoginTestWithPageObject extends BaseTest {
                     .enterTextIntoInputPassword(VALID_PASSWORD)
                     .clickOnButtonSignIn();
 
-        pageProvider.getHomePage().checkIsButtonSignOutVisible();
+        pageProvider.getHomePage().getHeaderElement().checkIsButtonSignOutVisible();
     }
 }
