+++ overAll/src/test/java/org/postTests/CreateNewPostTest.java
@@ -9,7 +9,7 @@ public class CreateNewPostTest extends BaseTest {
         pageProvider.getLoginPage()
                 .openLoginPageAndFillLoginFormWithValidCred()
                 .checkIsRedirectToHomePage()
-                .clickOnButtonCreatePost()
+                .getHeaderElement().clickOnButtonCreatePost()
                 .checkIsRedirectToCreateNewPostPage()
                 .enterTextIntoInputTitle("Title Taras")
                 .enterTextIntoInputBody("Body")
@@ -19,6 +19,8 @@ public class CreateNewPostTest extends BaseTest {
                 .checkTextInSuccessMessage("New post successfully created.")
 
         ;
+
+        pageProvider.getPostPage().getHeaderForUserElement().clickOnButtonMyProfile();
     }
 
 }
